+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ### 0.0.4 (unreleased)
+* Bump mikepenz/action-junit-report from 4 to 5 ([#17](https://github.com/f-cramer/jasperreports-gradle-plugin/pull/17))
 
 ### 0.0.3 (2024-11-12)
 * Create tmpDir before using it for compilation  ([#18](https://github.com/f-cramer/jasperreports-gradle-plugin/issues/18))