+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ### 0.0.4 (unreleased)
+* Bump com.gradle.plugin-publish from 1.2.2 to 1.3.0 ([#15](https://github.com/f-cramer/jasperreports-gradle-plugin/pull/15))
 
 ### 0.0.3 (2024-11-12)
 * Create tmpDir before using it for compilation  ([#18](https://github.com/f-cramer/jasperreports-gradle-plugin/issues/18))