+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ### 0.0.4 (unreleased)
+* Bump io.gitlab.arturbosch.detekt from 1.23.6 to 1.23.7 ([#14](https://github.com/f-cramer/jasperreports-gradle-plugin/pull/14))
 
 ### 0.0.3 (2024-11-12)
 * Create tmpDir before using it for compilation  ([#18](https://github.com/f-cramer/jasperreports-gradle-plugin/issues/18))