+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Bump `com.gradle.plugin-publish` from 1.2.2 to 1.3.0 ([#15](https://github.com/f-cramer/jasperreports-gradle-plugin/pull/15))
 - Bump `io.gitlab.arturbosch.detekt` from 1.23.6 to 1.23.7 ([#14](https://github.com/f-cramer/jasperreports-gradle-plugin/pull/14))
 - Bump `mikepenz/action-junit-report` from 4 to 5 ([#17](https://github.com/f-cramer/jasperreports-gradle-plugin/pull/17))
+- Bump `org.jlleitschuh.gradle.ktlint` from 12.1.1 to 12.1.2 ([#21](https://github.com/f-cramer/jasperreports-gradle-plugin/pull/21))
 
 ## 0.0.3 (2024-11-12)
 - Create tmpDir before using it for compilation  ([#18](https://github.com/f-cramer/jasperreports-gradle-plugin/issues/18))