+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Bump `io.gitlab.arturbosch.detekt` from 1.23.6 to 1.23.7 ([#14](https://github.com/f-cramer/jasperreports-gradle-plugin/pull/14))
 - Bump `mikepenz/action-junit-report` from 4 to 5 ([#17](https://github.com/f-cramer/jasperreports-gradle-plugin/pull/17))
 - Bump `net.researchgate.release` from 3.0.2 to 3.1.0 ([#22](https://github.com/f-cramer/jasperreports-gradle-plugin/pull/22))
+- Bump `org.jlleitschuh.gradle.ktlint` from 12.1.1 to 12.1.2 ([#21](https://github.com/f-cramer/jasperreports-gradle-plugin/pull/21))
 
 ## 0.0.3 (2024-11-12)
 - Create tmpDir before using it for compilation  ([#18](https://github.com/f-cramer/jasperreports-gradle-plugin/issues/18))